Выносит все константы в отдельный файл
