+++ src/components/main-page/main-page.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import {useParams} from 'react-router-dom';
+import {MAX_COUNT_OF_FILMS} from '../../const';
 import MainPageView from '../main-page/main-page-view';
-
-const MAX_COUNT_OF_FILMS = 8;
 
 const MainPage = () => {
