+++ src/const.js
@@ -1,8 +1,25 @@
 const RATINGS_COUNT = 10;
+const SECONDS_IN_HOUR = 60;
+const DELAY_TIME = 1000;
+const COUNT_OF_SIMULAR_FILMS = 4;
+const GENRES_COUNT = 9;
+const MAX_COUNT_OF_FILMS = 8;
+const TOGGLER_WIDTH = 17;
+
+const MouseEvents = {
+  MOVE: `mousemove`,
+  UP: `mouseup`
+};
+
 const ReviewLength = {
   MIN: 50,
   MAX: 400
 };
-const SECONDS_IN_HOUR = 60;
 
-export {RATINGS_COUNT, ReviewLength, SECONDS_IN_HOUR};
+const FavoriteStatus = {
+  FAVOURITE: 1,
+  NOT_FAVORITE: 0
+};
+
+
+export {RATINGS_COUNT, ReviewLength, SECONDS_IN_HOUR, DELAY_TIME, FavoriteStatus, COUNT_OF_SIMULAR_FILMS, GENRES_COUNT, MAX_COUNT_OF_FILMS, TOGGLER_WIDTH, MouseEvents};
